pyproject.toml
[tool.poetry]
name = "webdownloader-python-react"
version = "0.1.0"
description = "A put.io clone built with python, zeromq and reactjs"
authors = ["Gabriel Falcão <gabriel@nacaolivre.org>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.0"
pyzmq = "^18.0"
requests = "^2.21"
coloredlogs = "^10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
